moves custom folder inside the components changes port in package.json to 3001 refactors radio  button into custom component
